Só formatei o código e alterei para Long o id de categoria.

+++ src/main/java/br/edu/iftm/vitrino/repository/CategoriaRepository.java
@@ -1,10 +1,8 @@
 package br.edu.iftm.vitrino.repository;
 
-
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.edu.iftm.vitrino.entity.Categoria;
 
-public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {
+public interface CategoriaRepository extends JpaRepository<Categoria, Long> {
 }
